Extract colors editor.
